+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Return `[credit, debit]` from `DoubleEntry.transfer` ([#190]).
+
+[#190]: https://github.com/envato/double_entry/pull/190
+
 ## [2.0.0.beta5] - 2021-02-24
 
 ### Changed